container-video-embeddings/03-audio-video-workflow/requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
boto3
